refactor: 장바구니 리포지터리가 어플리케이션 생명주기에 따라가도록 변경
