+++ app/src/main/java/woowacourse/shopping/di/DiApplicationModule.kt
@@ -6,8 +6,11 @@ import com.re4rk.arkdi.Singleton
 import com.re4rk.arkdi.annotations.ContextType
 import com.re4rk.arkdi.annotations.ContextType.Type.APPLICATION
 import com.re4rk.arkdi.annotations.StorageType
+import com.re4rk.arkdi.annotations.StorageType.Type.DATABASE
 import com.re4rk.arkdi.annotations.StorageType.Type.IN_MEMORY
+import woowacourse.shopping.data.CartInDiskRepository
 import woowacourse.shopping.data.CartInMemoryRepository
+import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.ProductSampleRepository
 import woowacourse.shopping.data.ShoppingDatabase
 import woowacourse.shopping.repository.CartRepository
@@ -31,4 +34,15 @@ class DiApplicationModule(
     fun provideShoppingDatabase(
         @ContextType(APPLICATION) context: Context,
     ): ShoppingDatabase = ShoppingDatabase.getInstance(context)
+
+    @Singleton
+    fun provideCartProductDao(
+        shoppingDatabase: ShoppingDatabase,
+    ): CartProductDao = shoppingDatabase.cartProductDao()
+
+    @Singleton
+    @StorageType(DATABASE)
+    fun provideCartInDiskRepository(
+        cartProductDao: CartProductDao,
+    ): CartRepository = CartInDiskRepository(cartProductDao)
 }
